nptl: ia64: Fix Race conditions in pthread cancellation (BZ#12683)
This patch adds the ia64 modifications required for the BZ#12683 fix.

The syscall bridge uses the old brk 0x10000 instruction because by
using the vDSO gate the resulting PC value for an interrupted syscall
points to an address outside the expected markers in
__syscall_cancel_arch.  This is similar to i686 issue.

Also the __syscall_cancel_arch issues the 'break 0x100000' on its own
bundle, and __syscall_cancel_arch_end points to end of the previous one.
It requires an arch-specific ucontext_check_pc_boundary to check for the
ri value (embedded in the sc_ip by the kernel) to check if the syscall
had any side-effects.

Also, ia64 issues a sigcontext as third argument for sigaction handler
with SA_SIGINFO which requires an arch-specific ucontext_add_cancel.

Checked on ia64-linux-gnu.