linux: Add generic C syscall implementation
The hppa and riscv one are also removed.  The riscv specially is
abusing the internal interfaces by defining it with a different
signature.  If the idea is to provide an optimized version a better
alternative would be to provide an assembly one as other
architectures.

Checked on x86_64-linux-gnu and with build for the affected ABIs.