linux: Consolidate brk implementation
It removes all the arch-specific assembly implementation.  The
only outlier is alpha where its kernel ABI to return error
is to return the EINVAL as the returned value instead of address
lower than the requested one.

Also some ABIs exports an additional ___brk_addr symbol and to
handle it an internal HAVE_INTERNAL_BRK_ADDR_SYMBOL.

Checked on x86_64-linux-gnu and with build for the affected ABIs.