linux: Remove INTERNAL_SYSCALL_ERRNO
All architectures now returns the error code on the INTERNAL_SYSCALL
value and follow the Linux de-facto ABI (error are signaled with a
value between -4096UL and -1UL).  This make the INTERNAL_SYSCALL_ERRNO
superflous and some implementations already checks for either values
different than 0 or for an specific errno in the failure case.

Checked on x86_64-linux-gnu and powerpc64le-linux-gnu.