linux: Use generic __syscall_error for x86_64
The auto-generated syscalls issue the __syscall_error only for
static objects, so a arch specific syscall_error is added.  The
size of using extern __syscall_error is just slight better for
static case:

--- size-fstatat64.o.inline
+++ size-fstatat64.o.outline
@@ -1,2 +1,2 @@
    text           data     bss     dec     hex filename
-    100              0       0     100      64 io/fstatat64.o
+     94              0       0      94      5e io/fstatat64.o

For shared objects the inline __syscall_error does generate more
compact code:

--- sizes-x86_64-linux-gnu.inline
+++ sizes-x86_64-linux-gnu.outline
@@ -1,5 +1,5 @@
    text           data     bss     dec     hex filename
-1774028          20520   15896 1810444  1ba00c libc.so
- 167499           7280     392  175171   2ac43 elf/ld.so
-  99261           1640   16728  117629   1cb7d nptl/libpthread.so
-  24751           1188    2400   28339    6eb3 rt/librt.so
+1776260          20520   15896 1812676  1ba8c4 libc.so
+ 167758           7280     392  175430   2ad46 elf/ld.so
+ 100221           1640   16728  118589   1cf3d nptl/libpthread.so
+  24927           1188    2400   28515    6f63 rt/librt.so

A possible option might be to use the same strategy for both static
and PIC code, which allow to remove the arch-specific syscall_error.c
with a slight increase of code size in static library.

Checked on x86_64-linux-gnu.