libio: Assume _IO_MTSAFE_IO (BZ #29126)
It is already set by default on all supported architectures and it is
an expectation that stdio works on multi-threaded environments.  It
also fixes some internal usages of _IO_peekc, _IO_flockfile,
and_IO_funlockfile, where the lock is not taken because the file is
compiled without _IO_MTSAFE_IO (for instance malloc_stat from malloc.c).

Checked on x86_64-linux-gnu.