Move libio lock single-thread optimization to generic libc-lock
This patch moves the single thread stdio optimization (d2e04918833d9)
to the generic libc-lock.  It simplifies the support, since there is
no need to add per-function code to handle the single-thread case,
and removes the _IO_enable_locks (the SINGLE_THREAD_P uses the either
the TCB or the global variable which indicates wether the process
is multi-threaded).

As for current approach, once the process goes multithreaded the
single-thread is disabled (it won't be enable back if the process
become single-threaded).

It also handles the memory streams requirement (de895ddcd7fc), since
the SINGLE_THREAD_P already contains te information whether the
process is multithread (so there is no need to disable the optimization
because such stream are listed in _IO_list_all).

Checked on x86_64-linux-gnu, i686-linux-gnu, and aarch64-linux-gnu.