nptl: Do not use pthread set_tid_address as state synchronization (BZ #19951)
The use after free described in BZ#19951 is due the use of two different
PD fields, 'joinid' and 'cancelhandling', to describe the thread state.
And any state change potentially requires to check for both fields
atomically to potentially handle partial state (such as pthread_join()
with a cancellation handler and a joinstate rollback).

This patch uses a different PD member with 4 possible states (JOINABLE,
DETACHED, EXITING, and EXITED) to synchronize between pthread_create(),
pthread_join(), pthread_detached(), and pthread_exit():

  1. On pthread_create() the inital state is set either to JOINABLE or
     DETACHED depending of the pthread attribute used.

  2. On pthread_detach() call, a CAS is issued on the state.  If the CAS
     fails it means that thread is already detached (DETACHED) or is
     being terminated (EXITING).  For former an EINVAL is returned,
     while for latter pthread_detach() should be reponsible to join the
     thread (and deallocate the resources).

  3. On pthread_create() exit phase (reached either if the thread
     function has returned, pthread_exit() has being called, or
     cancellation handled has been acted upon) we issue a CAS on state
     to set to EXITING Mode.  If the thread is previous on DETACHED
     mode it will be the thread itself responsible to deallocate any
     resource, otherwise the threads needs to be joined.

  4. The clear_tid_field on 'clone' call is changed to set the new
     'state' field on thread exit (EXITED).  This state ins only
     reached on thread termination.

  4. The pthread_join() implementation is now simpler: the futex wait
     is done directly on thread state and there is no need to reset it
      in case of timeout (since the state is now set either by
      pthread_detach() or by the kernel on process termination).

The race condition on pthread_detach is avoided with only one atomic
operation on PD state: once the mode is set to THREAD_STATE_DETACHED
it is up to thread itself to deallocate its memory (done on the exit
phase at pthread_create()).

This change trigger an invalid C11 thread tests: it crates a thread,
which detaches itself, and after a timeout the creating thread checks
if the join fails.  The issue is once thrd_join() is called the thread
lifetime has already expired.  The test is changed so the slee is done
by the thread itself, so the creating thread will try to join a valid
thread.

Checked on x86_64-linux-gnu, i686-linux-gnu, aarch64-linux-gnu,
arm-linux-gnueabihf, and powerpc64-linux-gnu.