nptl: Do not use pthread set_tid_address as state synchronization (BZ #19951)
The use after free described in BZ#19951 is due the use of two
different PD fields, 'joinid' and 'cancelhandling', to describe the
thread state.  And the state change  potentially requires to check
for both fields atomically to either avoid partial state
(as pthread_join() tries to avoid with both a cancellation handler
and a joinstate rollback) or the race condition as described by the
bug.

This patch uses a different PD member with 4 possible states
(JOINABLE, DETACHED, EXITING, and EXITED) to synchronize between
pthread_create, pthread_join, pthread_detached, and pthread_exit:

  1. On pthread_create() the inital state is set either to JOINABLE or
     DETACHED depending of the pthread attribute used.

  2. On pthread_detach() call, a CAS is done on the state.  If the
     CAS fails it means that thread is already detached (DETACHED)
     or is being terminated (EXITING or EXITED).  For former an EINVAL
     is return, while for latter pthread_detach() should be reponsible
     to join the thread and deallocate the resources.

  3. On pthread_create() exit phase (reached either if the thread
     function has returned, pthread_exit() has being called, or
     cancellation handled has been acted upon) we issue a CAS on state
     to set to EXITING Mode.  If the thread is previous on DETACHED
     mode it will be the thread itself responsible to deallocate any
     resource, otherwise the threads needs to be joined.

     The EXITING mode exists to avoid a pthread_detach() to wrongly
     setup the mode while the thread calling the required exit
     functions, but has not while deallocate the stack.

  4. The clear_tid_field on 'clone' call is changed to set the new
     'state' field on thread_exit (EXITED).

  4. The pthread_join() implementation is now much simpler: the futex
     wait is done directly on thread state and there is no need to
     reset it in case of TIMEOUT (since the state is now set either by
     pthread_detached or by kernel on thread termination).

The race condition on pthread_detach is avoided with only one atomic
operation on PD state: once the mode is set to THREAD_STATE_DETACHED
it is up to thread itself to deallocate its memory (done on the exit
phase at pthread_create()).

Checked on x86_64-linux-gnu, i686-linux-gnu, aarch64-linux-gnu,
arm-linux-gnueabihf, and powerpc64-linux-gnu.