nptl: Handle robust PI mutexes for !__ASSUME_SET_ROBUST_LIST
The robust PI mutexes are signaled by setting the LSB bit to 1, so
the code requires to take this consideration before access the
__pthread_mutex_s.

The code is also simplifies (the initialization code is not really
required, pd->robust_head.list and pd->robust_list.__next are
essentially the same for !__PTHREAD_MUTEX_HAVE_PREV), the futex wake
is optimized to be issued only when required, and the the futex shared
bit is set only when required.

Checked on a build for m68k-linux-gnu.