stdio: Replace _IO_flockfile with _IO_acquire_lock or flockfile
Using _IO_flockfile internally forces the stream be the always locked,
since _IO_FLAGS2_NEED_LOCK is not cleared once it set.  It also
prevents the single-thread optimization on further stdio functions
and does not release the lock in case of exceptions.  This is not
ideal for libio code (clearerr, feof, ferror, and internal link
functions).

For other usages replacing _IO_flockfile/_IO_funlockfile with
__flockfile/__funlock is avoids bleeding libio implementation details.

Checked on x86_64-linux-gnu.