csu: Fix static when PIE is enabled but static PIE is not supported (BZ #29575)
On architectures where static PIE is not supported, both libc and
static binaries are built still with -fPIE and -DPIC.  It leads to
wrong code for some architectures (for instance sparc).

This patch only enables -DPIC for .o files iff static-pie is also
supported.  Since the same rule is used for both libc.a objects and
installed binaries, the -fno-pie is added if static-pie is not
supported (with a strategy similar to how stack protector is
disabled for some architectures).

Checked on x86_64-linux-gnu (with and without --disable-static-pie)
and I also checked on sparc64-linux-gnu to verify that static binaries
are not broken and both binaries and tests continue to be built
for PIE.